Create SPARSE-TABLE
This is short implementation of sparse table which is generally used to calculate range queries BUT NOT UPDATES
Sparse Table is a data structure, that allows answering range queries. It can answer most range queries in O(logn), but its true power is answering range minimum queries (or equivalent range maximum queries). For those queries it can compute the answer in O(1) time.
The time complexity of the precomputation is O(NlogN).